requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
gunicorn==21.2.0